+++ src/main/java/com/onevizion/uitest/api/helper/Qs.java
@@ -164,11 +164,11 @@ public class Qs {
         fillBooleanQsValue(gridIdx, search);
     }
 
-    public void searchValueAndCheck(String fieldName, String search, Long rowsCnt, List<String> expectedValues) {
+    public void searchValueAndCheck(String fieldName, String search, int rowsCnt, List<String> expectedValues) {
         searchValueAndCheck(AbstractSeleniumCore.getGridIdx(), fieldName, search, rowsCnt, expectedValues);
     }
 
-    public void searchValueAndCheck(Long gridIdx, String fieldName, String search, Long rowsCnt, List<String> expectedValues) {
+    public void searchValueAndCheck(Long gridIdx, String fieldName, String search, int rowsCnt, List<String> expectedValues) {
         int columnIndex = js.getColumnIndexByLabel(gridIdx, fieldName);
 
         searchValue(gridIdx, fieldName, search);
@@ -190,11 +190,11 @@ public class Qs {
         clickClearQs(gridIdx);
     }
 
-    public void searchBooleanValueAndCheck(String fieldName, String search, Long rowsCnt, List<String> expectedValues) {
+    public void searchBooleanValueAndCheck(String fieldName, String search, int rowsCnt, List<String> expectedValues) {
         searchBooleanValueAndCheck(AbstractSeleniumCore.getGridIdx(), fieldName, search, rowsCnt, expectedValues);
     }
 
-    public void searchBooleanValueAndCheck(Long gridIdx, String fieldName, String search, Long rowsCnt, List<String> expectedValues) {
+    public void searchBooleanValueAndCheck(Long gridIdx, String fieldName, String search, int rowsCnt, List<String> expectedValues) {
         int columnIndex = js.getColumnIndexByLabel(gridIdx, fieldName);
 
         searchBooleanValue(gridIdx, fieldName, search);
@@ -232,7 +232,7 @@ public class Qs {
      * checkUserpageGridTextColumnEquals and checkAdminpageGridTextColumnEquals similar as userpageFilter.checkGridTextColumnEquals
      */
     private void checkUserpageGridTextColumnEquals(Long gridId, int columnIndex, List<String> expectedValues) {
-        Long rowsCnt = grid.getGridRowsCount(gridId);
+        int rowsCnt = grid.getGridRowsCount(gridId);
         @SuppressWarnings("unchecked")
         List<String> vals = (List<String>) js.getGridCellsValuesTxtForColumnByColIndex(gridId, rowsCnt, columnIndex);
 
@@ -259,7 +259,7 @@ public class Qs {
     private void checkToGridTextColumnEquals(Long gridId, int columnIndex, List<String> expectedValues) {
         int differenceInRows = js.getToGridDatePairCount(gridId);
 
-        Long rowsCnt = grid.getTOGridRowsCount(gridId);
+        int rowsCnt = grid.getTOGridRowsCount(gridId);
         @SuppressWarnings("unchecked")
         List<String> vals = (List<String>) js.getGridCellsValuesTxtForColumnByColIndex(gridId, rowsCnt * differenceInRows, columnIndex);
 
@@ -284,7 +284,7 @@ public class Qs {
      * checkUserpageGridTextColumnEquals and checkAdminpageGridTextColumnEquals similar as userpageFilter.checkGridTextColumnEquals
      */
     private void checkAdminpageGridTextColumnEquals(Long gridId, int columnIndex, List<String> expectedValues) {
-        Long rowsCnt = grid.getGridRowsCount(gridId);
+        int rowsCnt = grid.getGridRowsCount(gridId);
         @SuppressWarnings("unchecked")
         List<String> vals = (List<String>) js.getGridCellsValuesForColumnByColIndexNew(gridId, rowsCnt, columnIndex);
