+++ src/main/java/com/onevizion/uitest/api/helper/entity/EntityDynamicVtableValue.java
@@ -131,7 +131,7 @@ public class EntityDynamicVtableValue {
         window.closeModal(By.id(AbstractSeleniumCore.BUTTON_CANCEL_ID_BASE));
     }
 
-    public void testInGrid(Long gridId, Long rowIndex, DynamicVtableValue dynamicVtableValue) {
+    public void testInGrid(Long gridId, int rowIndex, DynamicVtableValue dynamicVtableValue) {
         Map<Integer, String> gridVals = new HashMap<>();
 
         gridVals.put(js.getColumnIndexByLabel(gridId, "Value"), dynamicVtableValue.getValue());
